game/...: rename isTouching -> hasAnyInside
the new name more accurately reflects what the method does. it is mean
to be read 'does this sprite have any of the s sprite inside of it?'
realizing what the method really does prevents a 'bug' where the method
was called with a this was a smaller sprite than s.