cbg/catan_board: fix infinite loop issue
when the desert tile was the last tile to be placed, the `labels` array
was all true, since all the labels were used up, but the `tiles` array
was all true except for the last element, which represents the desert
tile. then, we got caught trying to get an index of a false in an array
of only true.

to mitigate, only get a random label index when we actually need to
(that is, when the type of the tile is NOT desert).