Fib about the date
Never had two posts in one day before, but now there's the problem of
the sort order not actually being the order in which I wrote the posts.
This just kicks the problem down the road by fibbing and saying I wrote
the post tomorrow. It's late here, so it's almost not a lie.